requirements.txt
anthropic>=0.21.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.0.0
pillow-heif>=0.13.0